refactor: update test case
