+++ static/js/redux/__tests__/selectors.test.jsx
@@ -1,4 +1,4 @@
-import { getSectionTypeToSections, getMaxEndHour } from "../state/entities_reducer";
+import { getSectionTypeToSections, getMaxEndHour } from "../state/slices/entitiesSlice";
 
 describe("course selectors", () => {
   describe("section type to sections selector", () => {
